Cleanup unused code and reorganize files in packages
