+++ src/main/java/it/unibz/inf/ontouml/vp/controllers/ApplyAssociationStereotypeId.java
@@ -1,10 +1,8 @@
-package it.unibz.inf.ontouml.vp.utils;
+package it.unibz.inf.ontouml.vp.controllers;
 
+import it.unibz.inf.ontouml.vp.utils.Stereotype;
 import java.util.Arrays;
 import java.util.HashMap;
-import java.util.Map;
-import java.util.Set;
-import java.util.stream.Collectors;
 
 public enum ApplyAssociationStereotypeId {
