+++ src/main/java/it/unibz/inf/ontouml/vp/controllers/ApplyAttributeStereotypeId.java
@@ -1,5 +1,6 @@
-package it.unibz.inf.ontouml.vp.utils;
+package it.unibz.inf.ontouml.vp.controllers;
 
+import it.unibz.inf.ontouml.vp.utils.Stereotype;
 import java.util.Arrays;
 import java.util.HashMap;
