+++ src/main/java/it/unibz/inf/ontouml/vp/controllers/ApplyClassStereotypeMenuManager.java
@@ -1,9 +1,12 @@
-package it.unibz.inf.ontouml.vp.utils;
+package it.unibz.inf.ontouml.vp.controllers;
 
 import com.vp.plugin.action.VPAction;
 import com.vp.plugin.action.VPContext;
 import com.vp.plugin.model.IClass;
 import it.unibz.inf.ontouml.vp.model.uml.Class;
+import it.unibz.inf.ontouml.vp.utils.OntoUMLConstraintsManager;
+import it.unibz.inf.ontouml.vp.utils.StereotypesManager;
+import it.unibz.inf.ontouml.vp.utils.VPContextUtils;
 
 public class ApplyClassStereotypeMenuManager extends ApplyStereotypeMenuManager {
